Fix crash when not passing environment variables
`action='append'` doesn't change the default, so we were trying to
iterate over None. We can't set the default to the empty list, because
`action='append'` will append to the actual list object that's set as
the default in that case. That wouldn't cause a noticeable effect
today, but it's not what we want and it would be a likely source of
future bugs.